Begin the RobotController architecture
WPILib works well for running the robot, but it's not very testable
since it depends so much on the actual roboRIO hardware. To improve the
testability of the robot, this begins the process of abstracting the
robot into a controller vs the robot actuators.

Here, IRobot represents an "actual robot," the pieces of code that run
actuators and the robot control life-cycle (game mode changes and loop).
The PhysicalRobot is an implementation of that concept that's intended
to control our actual prototype robot. It's a subclass of the WPILib
IterativeRobot class in order to provide the complex communication with
the field that's necessary. IRobot has methods to set the speeds of the
various motors on our robot.

IRobotController represents the logic of the robot. It's expected that
the IRobot will delegate to the IRobotController in order to perform
periodic updates and initialization.